% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateUID.R
\name{generateUID}
\alias{generateUID}
\title{Generate Unique Identifier}
\usage{
generateUID(codeSize = 11)
}
\arguments{
\item{codeSize}{length of UID, default = 11}
}
\description{
Generate Unique Identifier
}
\details{
used for DATIM, author - Jason Pickering
}
