% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_msd.R
\name{mask_msd}
\alias{mask_msd}
\title{Mask MSD, aka Training Dataset}
\usage{
mask_msd(msd_filepath, psnuuids)
}
\arguments{
\item{msd_filepath}{full file path to the MSD (PSNUxIM) (.txt)}

\item{psnuuids}{list of PSNU UIDs to select (n must equal 15)}
}
\description{
Mask MSD, aka Training Dataset
}
\examples{
 \dontrun{
  #filepath for MSD (.txt)
    msd_filepath <- "~/ICPI/Data/MER_Structured_Dataset_PSNU_IM_FY17-18_20181115_v1_2.txt"
  #supply list of PSNU uids to use
    psnuuid_list <- c("QlWUt1rBsEo", "GfzVv4oeclF", "PNgOI7VPe98", "TWpkDEvK6si", "aqktQTp0wHa", 
                      "XFdvhW8Ga1S", "sVwvt4bYesp", "EJDcY4F1rsj", "nXEQ97b2YHJ", "DfXQBOLWwbZ", 
                      "aC69ENcI2hU", "PJpAerXQjZ7", "PAj75tgxkIU", "Wru5kJQ36GT", "HB75Phs4wZL")
  #generate training dataset
    gen_training_planets(msd_filepath, psnuuid_list) 
  }
}
