% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planets_geo.R
\docType{data}
\name{planets_geo}
\alias{planets_geo}
\title{Geography for the Masked Dataset}
\format{A dataframe with 15 rows and 9 variables:
\describe{
  \item{Region_mw}{hieararchy one level below global}
  \item{RegionUID_mw}{the unique ID for Region}
  \item{OperatingUnit_mw}{the operating unit/country}
  \item{OperatingUnitUID_mw}{the unique ID for the Operating Unit}
  \item{CountryName_mw}{the country name, ie for regional missions}
  \item{SNU1_mw}{sub national unit 1, 1 level below OU}
  \item{SNU1Uid_mw}{unique ID for the SNU1}
  \item{PSNU1_mw}{sub national level targeting occurs at}
  \item{PSNUuid_mw}{unique ID for the PSNU}
}}
\source{
Put together by Abe Agedew (CDC/ICPI)
}
\usage{
planets_geo
}
\description{
A dataset containing MER Structured Dataset
hieararchy for 15 Priority Sub Natioaal Units (PSNUs).
All data have an underscore of _mw to designate they 
fall in the Milky Way (they duplicate the columns
in the MSD they will replace).
}
\keyword{datasets}
