% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planets_primepartners.R
\docType{data}
\name{planets_primepartners}
\alias{planets_primepartners}
\title{List of Prime Partners and Mechanism Names}
\format{A data frame with 354 rows and 2 variables.
\describe{
  \item{primepartner_mw}{Masked names for prime partners}
  \item{implementingmechanismname_mw}{Masked names for mechanisms}
}}
\source{
List of proper names of stars, Wikipedia 
  \url{https://en.wikipedia.org/wiki/List_of_proper_names_of_stars}
}
\usage{
planets_primepartners
}
\description{
A dataset to mask partner and mechanism names.
}
\keyword{datasets}
