% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planets_site_types.R
\docType{data}
\name{planets_site_types}
\alias{planets_site_types}
\title{List of site types}
\format{A vector of 7}
\usage{
planets_site_types
}
\description{
A vector of facility types to help mask site name.
}
\keyword{datasets}
