% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planets_sites.R
\docType{data}
\name{planets_sites}
\alias{planets_sites}
\title{List of astronomy terms for site names}
\format{A vector of 266}
\source{
Glossary of astronomy, Wikipedia 
  \url{https://en.wikipedia.org/wiki/Glossary_of_astronomy}
}
\usage{
planets_sites
}
\description{
A vector of astronomy terms to mask site names
}
\keyword{datasets}
