% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_msd.R
\name{mask_msd}
\alias{mask_msd}
\title{Mask MSD, aka Training Dataset}
\usage{
mask_msd(msd_filepath, ...)
}
\arguments{
\item{msd_filepath}{full file path to the MSD (PSNUxIM) (.txt)}

\item{...}{PSNUS to select (15)}
}
\description{
Mask MSD, aka Training Dataset
}
\examples{
 \dontrun{
  #filepath for MSD (.txt)
    msd_filepath <- "~/ICPI/Data/MER_Structured_Dataset_PSNU_IM_FY17-18_20181115_v1_2.txt"
  #genreate training dataset
    gen_training_planets(msd_filepath) 
  }
}
